Wipe out old approach to cloaking
While it would be nice to keep all of the variant-specific stuff
localized to just the descendant of `Game` (as has been possible for
most other variants), cloaked pieces need to know about each other, so
to speak, such that methods like `Piece#vision` and `Piece#moves` return
the right results, particularly when determining checks.

`CloakDaggerGame#legal_moves_for` isn't the right place to try to do this.